ImageHolder, InfoDisplayExtra, InfoDiplay implementation.
And also, did a lot with coroutines this time, now the webRequest will be resfreshed every 20 secs or some of that shit. Implemented also the python graphic, basically converted from a .JSON to base64, and therefore to the sprite/image/png. Implemented coroutines for the graphic. Changed a little the blooming effect so they will be aligned and anti-alised a little bit and comforted for VR camera angle. Managing to do the final task and parse the webRequest that was outside the .JSON range in some way, now it is correctly displayable in current method of the StartCoroutine fucntion. Also changed Inumerator in GetImagePost in a way that the case will be responsible to launch data foreach. That's it i guess.

Post scriptum - Casimiro, respond if ya see that message bellow lol.